[nxp][platform][common] Move OTA related files to a common folder
OTAFirmwareProcessor and OTAHooks could be reused by other platforms
as long as the SDK is the same, so moving them from k32w1 folder.
